fix: pr结果修改
- 修改了view文件的数量
- 要求了用户的唯一性
- migration文件清理
