--- apps/account/Views/request_mail_verify_view.py
@@ -1,148 +1,0 @@
-import secrets
-import string
-import random
-from apps.account.models import Account, LoginLog, EmailAuthCodeChoice, AccountEmailAuthCode, AccountEmailAuthLog
-from libs.boost.parser import Argument, JsonParser
-from libs.boost.http import JsonResponse
-from libs.boost.http import HttpStatus
-from libs.email.netease import MailServer
-from const.error import ErrorType
-from django.views.generic import View
-from django.conf import settings
-from django.utils import timezone
-from datetime import timedelta
-from django.db import transaction
-from utils.utils import generate_random_str
-from django.conf import settings
-
-
-class RequestMailVerifyView(View):
-
-    __EMAIL_SUBJECT__ = "数字化工具库-验证码"
-
-    def get(self, request) -> JsonResponse:
-        form, err = JsonParser(
-            Argument('email', data_type =str),
-            Argument('for_what', data_type=EmailAuthCodeChoice)
-            # Argument('for_what', data_type=[1,2,3],())
-            ).parse(request.body)
-        
-        if (form.for_what == EmailAuthCodeChoice.REGISTER):
-            return self._regist_new_account(form.email)
-        elif (form.for_what == EmailAuthCodeChoice.PASSWORD):
-            return self._change_password(form.email)
-        else:
-            return JsonResponse(error_type=ErrorType.REQUEST_ILLEGAL)
-
-
-    def _change_password(self, email: str) -> JsonResponse:
-        '''
-        重置密码的发送验证码逻辑
-        '''
-        account: Account = Account.objects.filter(email=email).first()
-        if account is None:
-            return JsonResponse(error_type=ErrorType.ACCOUNT_NOT_EXIST)
-        code = generate_random_str(5, False)
-
-        email_auth_code = AccountEmailAuthCode.objects.filter(
-            email=email,
-            for_what=EmailAuthCodeChoice.PASSWORD
-            ).order_by('-id').first()
-        
-        if email_auth_code is None:
-            # 不存在则创建
-            email_auth_code = AccountEmailAuthCode(
-                email=email,
-                code=code, 
-                is_valid=True,
-                expired = timezone.now() + timedelta(minutes=settings.VERIFY_CODE_EXPIRED),
-                for_what=EmailAuthCodeChoice.PASSWORD
-                )
-        else:
-            email_auth_code.code = code
-            email_auth_code.is_valid = True
-            email_auth_code.expired = timezone.now() + timedelta(minutes=settings.VERIFY_CODE_EXPIRED)
-        
-        # 发送验证码
-        self.__send_verify_code(email_auth_code.code, email_auth_code.email)
-        # 创建发送日志
-        email_auth_log = AccountEmailAuthLog(
-            email=email,
-            code=code,
-            is_success=False,
-            for_what=EmailAuthCodeChoice.PASSWORD
-            )
-        email_auth_code.save()
-        email_auth_log.save()
-        return JsonResponse(data='验证码已发送', status_code=HttpStatus.HTTP_200_OK)
-
-
-    def _regist_new_account(self, email: str) -> JsonResponse:
-        '''
-        注册新用户的发送验证码逻辑
-        '''
-        is_account_exist: Account = Account.objects.filter(email=email).exists()
-
-        if is_account_exist:
-            return JsonResponse(error_type=ErrorType.ACCOUNT_EXIST)
-        
-        code = generate_random_str(5, False)
-        
-        email_auth_code = AccountEmailAuthCode.objects.filter(email=email, for_what=EmailAuthCodeChoice.REGISTER).order_by('-id').first()
-        if email_auth_code is None:
-        # 不存在则创建
-            email_auth_code = AccountEmailAuthCode(
-                email=email,
-                code=code, 
-                is_valid=True,
-                expired = timezone.now() + timedelta(minutes=settings.VERIFY_CODE_EXPIRED),
-                for_what=EmailAuthCodeChoice.REGISTER
-                )
-        else:
-            email_auth_code.code = code
-            email_auth_code.is_valid = True
-            email_auth_code.expired = timezone.now() + timedelta(minutes=settings.VERIFY_CODE_EXPIRED)
-        # 创建发送日志
-        email_auth_log = AccountEmailAuthLog(
-            email=email,
-            code=code,
-            is_success=False,
-            for_what=EmailAuthCodeChoice.REGISTER
-        )
-        email_auth_code.save()
-        email_auth_log.save()
-        
-        # 发送验证码
-        self.__send_verify_code(email_auth_code.code, email_auth_code.email)
-        return JsonResponse(data='验证码已发送', status_code=HttpStatus.HTTP_200_OK)
-    
-    def __send_verify_code(self, verify_code: str, email: str) -> None:
-        '''
-        向指定邮箱发送验证码
-        '''
-        mail_server = MailServer()
-        mail_server.login()
-        mail_server.send([email], self.__EMAIL_SUBJECT__, self.__generate_mail_content(email, verify_code), [])
-        mail_server.quit()
-
-
-    def __generate_mail_content(self, email: str, verify_code: str) -> str:
-        '''
-        生成邮件内容
-        '''
-        content =  """
-            <!DOCTYPE html>
-            <html>
-            <head>
-                <meta charset="UTF-8">
-                <title>Email Verification Code</title>
-            </head>
-            <body>
-                <h2>验证码通知</h2>
-                <p>您已选择 <u>{email}</u> 作为您的邮箱账户，为验证此电子邮箱属于您，请在您的邮箱验证界面输入下方6位验证码：</p>
-                <p id="verification-code">{code}</p>
-                <p>此验证码将于2个小时后失效，请您尽快完成验证。</p>
-            </body>
-            </html>
-            """.format(email=email, code=verify_code)
-        return content
